enhancement/9 Split App module
